+++ src/main/java/com/SUG/FLORA/model/IntDomain.java
@@ -14,32 +14,32 @@ import lombok.Data;
 @MappedSuperclass
 @Data
 public class IntDomain {
-	
+
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-		
+	private long id;
+
 	@Basic
 	@Column(name = "creation_date")
 	private LocalDateTime creationDate = LocalDateTime.now();
-	
+
 	@Basic
 	private boolean deleted = false;
-	
+
 	@Basic
 	@Column(name = "last_update")
 	private LocalDateTime lastUpdate = LocalDateTime.now();
-	
+
 	@Basic
-    @Column
-    private LocalDateTime deletedDate;
+	@Column
+	private LocalDateTime deletedDate;
 
 	public void copyDomainOfIntDomainDTO(IntDomainDTO dto) {
 		setId(dto.getId());
-        setCreationDate(dto.getCreationDate());
-        setDeleted(dto.isDeleted());
-        setDeletedDate(dto.getDeletedDate());
-        setLastUpdate(dto.getLastUpdate());
+		setCreationDate(dto.getCreationDate());
+		setDeleted(dto.isDeleted());
+		setDeletedDate(dto.getDeletedDate());
+		setLastUpdate(dto.getLastUpdate());
 	}
-	
+
 }
